Implement `DrugSummaryUi` on `DrugSummaryView`
