+++ app/src/main/java/org/simple/clinic/summary/prescribeddrugs/DrugSummaryViewInjector.kt
@@ -1,0 +1,5 @@
+package org.simple.clinic.summary.prescribeddrugs
+
+interface DrugSummaryViewInjector {
+  fun inject(target: DrugSummaryView)
+}
